refactor(chat): move joinBaseUrl helper to utils
joinBaseUrl is a generic URL utility with no model-fetcher specifics, so it
belongs in utils/ (imported via @/utils/base-url) rather than under the chat
routes. No behavior change.
