+++ platform/backend/src/routes/chat/model-fetchers/vllm.ts
@@ -1,6 +1,6 @@
 import config from "@/config";
 import logger from "@/logging";
-import { joinBaseUrl } from "./base-url";
+import { joinBaseUrl } from "@/utils/base-url";
 import { type ModelInfo, PLACEHOLDER_BEARER_TOKEN } from "./types";
 
 export async function fetchVllmModels(
