+++ platform/backend/src/routes/chat/model-fetchers/xai.ts
@@ -1,6 +1,6 @@
 import config from "@/config";
 import type { OpenAi } from "@/types";
-import { joinBaseUrl } from "./base-url";
+import { joinBaseUrl } from "@/utils/base-url";
 import { fetchModelsWithBearerAuth } from "./openai-compatible";
 import type { ModelInfo } from "./types";
